adjust code style

+++ lib/dumper.cpp
@@ -79,24 +79,21 @@ void Dumper::writeDataBinary(unsigned len, double *yData, double *xData){
 	if (yData == nullptr)
 		return;
 
-	unsigned dataLen = len * sizeof(yData[0]);
-	long unsigned int bytesWritten = write(socketFd, &dataLen, sizeof(dataLen));
-	if (bytesWritten != sizeof(len))
-	{
+	size_t dataLen = len * sizeof(yData[0]);
+	ssize_t bytesWritten = write(socketFd, &dataLen, sizeof(dataLen));
+	if ((size_t) bytesWritten != sizeof(len)) {
 		logger->warn("Could not send all content to socket {}", socketName);
 		warningCounter++;
 	}
 
 	bytesWritten = write(socketFd, "d000", 4);
-	if (bytesWritten != 4)
-	{
+	if (bytesWritten != 4) {
 		logger->warn("Could not send all content to socket {}", socketName);
 		warningCounter++;
 	}
 
-	bytesWritten =  write(socketFd, yData, dataLen );
-	if (bytesWritten != dataLen && (!supressRepeatedWarning || warningCounter <1 ))
-	{
+	bytesWritten = write(socketFd, yData, dataLen );
+	if (bytesWritten != (ssize_t) dataLen && (!supressRepeatedWarning || warningCounter <1 )) {
 		logger->warn("Could not send all content to socket {}", socketName);
 		warningCounter++;
 	}
@@ -104,21 +101,19 @@ void Dumper::writeDataBinary(unsigned len, double *yData, double *xData){
 
 void Dumper::writeDataCSV(unsigned len, double *yData, double *xData)
 {
-	for (unsigned i = 0; i<len; i++)
-	{
+	for (unsigned i = 0; i < len; i++) {
 		std::stringstream ss;
 
 		ss << yData[i];
 
-		if(xData != nullptr)
+		if (xData != nullptr)
 			ss << ";" << xData[i];
 
 		ss << std::endl;
 
 		auto str = ss.str();
 		auto bytesWritten =  write(socketFd, str.c_str(), str.length());
-		if ((long unsigned int) bytesWritten != str.length() && (!supressRepeatedWarning || warningCounter <1 ))
-		{
+		if ((size_t) bytesWritten != str.length() && (!supressRepeatedWarning || warningCounter < 1)) {
 			logger->warn("Could not send all content to socket {}", socketName);
 			warningCounter++;
 		}
